move function out of extension.ts. Use in BaseCommand function.
