+++ src/extension.ts
@@ -30,10 +30,11 @@ import {
   uninstall,
   cleanup
 } from "./one-click/install";
-
 import { getChildProcessPath, getChildProcessProsToolchainPath  } from "./one-click/path";
 import { TextDecoder, TextEncoder } from "util";
 import { Logger } from "./logger";
+import { get_cwd_is_pros } from "./workspace";
+
 let analytics: Analytics;
 
 export var system: string;
@@ -90,49 +91,6 @@ export const getProsTerminal = async (
     env: process.env,
   });
 };
-
-/*!
-  * @brief This function returns the current working directory (vscode.Uri) and if it is a pros project (boolean)
-  * @return [vscode.Uri, boolean]
-  * 
-  * @details This function firstly looks at the workspace for the active text editor. If there is no active text editor, it will use the 0th index of the workspace folders. 
-  * If there are no workspace folders, it will throw an error. If the workspace folder is a pros project, it will return the workspace folder and true. 
-  * If the workspace folder is not a pros project, it will return the workspace folder and false.
-  * 
-  */
-
-export const get_cwd_is_pros = async (): Promise<[vscode.Uri, boolean]> => {
-  //output the 0th workspace folder
-
-  let active = vscode.window.activeTextEditor?.document.uri ?? undefined;
-  let active_dir = undefined;
-
-  const filename_search = "project.pros";
-  let is_pros_project = true;
-
-
-  if(active !== undefined) {
-    console.log(`active: ${active}`);
-    active_dir = vscode.workspace.getWorkspaceFolder(active)?.uri;
-    console.log(`workspace folder: ${active_dir}`);
-  } else if(vscode.workspace.workspaceFolders !== undefined && vscode.workspace.workspaceFolders !== null) {
-      active_dir = vscode.workspace.workspaceFolders[0].uri;
-  } 
-
-  if(active_dir === undefined || active_dir === null) {
-    throw(new Error("No workspace folder found"));
-  }
-
-
-  // use fs to check if the active directory contains a project.pros file
-  try {
-    await vscode.workspace.fs.stat(vscode.Uri.joinPath(active_dir, filename_search));
-  } catch (err) {
-    is_pros_project = false;
-  }
-  return [active_dir, is_pros_project];
-
-}
 
 export function activate(context: vscode.ExtensionContext) {
   vscode.window.showInformationMessage("PROS extension activated");
@@ -460,6 +418,7 @@ export function deactivate() {
 async function workspaceContainsProjectPros(): Promise<boolean> {
   return ((await get_cwd_is_pros())[1]);
 }
+
 //This code calls prosProjects and allows user to choose which pros project to work on
 async function chooseProject() {
   if (
